any type file upload and view and post total modal crud with forum comment done

+++ app/Http/Controllers/CategoryController.php
@@ -33,15 +33,15 @@ class CategoryController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         //Validation
-         $this->validate($request, [
+        $this->validate($request, [
             'name' => 'required|max:255'
-         ]);
+        ]);
 
         Category::create($request->all());
 
@@ -52,7 +52,7 @@ class CategoryController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Category  $category
+     * @param \App\Category $category
      * @return \Illuminate\Http\Response
      */
     public function show(Category $category)
@@ -63,19 +63,19 @@ class CategoryController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Category  $category
+     * @param \App\Category $category
      * @return \Illuminate\Http\Response
      */
     public function edit(Category $category)
     {
-       return view('admin.category.edit', compact('category'));
+        return view('admin.category.edit', compact('category'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Category  $category
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Category $category
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Category $category)
@@ -90,15 +90,21 @@ class CategoryController extends Controller
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Category  $category
-     * @return \Illuminate\Http\Response
+     * @param Category $category
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Exception
      */
     public function destroy(Category $category)
     {
-       $category->delete();
-        return redirect('/categories')->with('status', 'Deleted successful');
+        if ($category->posts->count() == 0) {
+            $category->delete();
+            return redirect('/categories')->with('status', 'Deleted successful');
+        } else {
+            return redirect('/categories')->with('status', 'Deleted unsuccessful. This category belongs to a Post');
+
+        }
+
+
     }
 
     /**
